examples/python-webhooks/requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
miro_api==2.2.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.19
Werkzeug==3.0.3
